+++ src/mainwid.cpp
@@ -37,7 +37,7 @@ const char mainwid::m_seq_to_char[c_seqs_in_set] =
 
 // Constructor
 
-mainwid::mainwid( perform *a_p  ): DrawingArea(), seqmenu( a_p )
+mainwid::mainwid( perform *a_p, mainwnd *a_main  ): DrawingArea(), seqmenu( a_p, a_main )
 {
     using namespace Menu_Helpers;