+++ src/midifile.h
@@ -53,7 +53,7 @@ class midifile
     ~midifile();
 
     bool parse( perform *a_perf, int a_screen_set );
-    bool write( perform *a_perf );
+    bool write( perform *a_perf, int a_screen_set, int a_sequence);
 
 };